#21 gRPC common
- core and protoc hierarchy change
