+++ pom.xml
@@ -219,6 +219,7 @@
     	<module>model</module>
         <module>common</module>
         <module>sample</module>
+        <module>grpc</module>
 <!--
         <module>sample-quarkus</module>
 -->
