+++ sample/sample-grpc-server-service/src/main/java/hu/icellmobilsoft/sampler/sample/grpc/server/service/error/mapper/BaseExceptionMapper.java
@@ -33,7 +33,7 @@ import hu.icellmobilsoft.coffee.dto.exception.BusinessException;
 import hu.icellmobilsoft.coffee.dto.exception.DtoConversionException;
 import hu.icellmobilsoft.coffee.dto.exception.ServiceUnavailableException;
 import hu.icellmobilsoft.coffee.se.logging.Logger;
-import hu.icellmobilsoft.sampler.common.grpc.core.exception.ExceptionMapper;
+import hu.icellmobilsoft.sampler.grpc.core.exception.ExceptionMapper;
 
 /**
  * Implementation of {@link ExceptionMapper} that maps {@link BaseException} to gRPC {@link Status}.
