+++ sample/sample-grpc-server-service/src/main/java/hu/icellmobilsoft/sampler/sample/grpc/server/service/error/mapper/GeneralExceptionMapper.java
@@ -29,7 +29,7 @@ import com.google.rpc.Status;
 
 import hu.icellmobilsoft.coffee.dto.exception.enums.CoffeeFaultType;
 import hu.icellmobilsoft.coffee.se.logging.Logger;
-import hu.icellmobilsoft.sampler.common.grpc.core.exception.ExceptionMapper;
+import hu.icellmobilsoft.sampler.grpc.core.exception.ExceptionMapper;
 
 /**
  * Implementation of {@link ExceptionMapper} that maps general Exceptions to gRPC {@link Status}. If the exception is not recognized, it maps it to
